challenge7/requirements.txt
semantic-kernel
python-dotenv